add error handler middleware

+++ src/controllers/userController.ts
@@ -1,4 +1,5 @@
 import { NextFunction, Request, Response } from "express";
+import ApiError from "../error/ApiError";
 import { User } from "../models/models";
 
 class userContoller {
@@ -7,16 +8,23 @@ class userContoller {
         return res.json(users);
     }
 
-    async addUser(req: Request, res: Response) {
+    async addUser(req: Request, res: Response, next: NextFunction) {
         const { name } = req.body;
+        if (!name) {
+            return next(ApiError.badRequest("not properties name"));
+        }
         const list = await User.create({
             name
         });
         res.json(list);
     }
 
-    async deleteUser(req: Request, res: Response) {
+    async deleteUser(req: Request, res: Response, next: NextFunction) {
         const { id } = req.body;
+
+        if (!id) {
+            return next(ApiError.badRequest("not properties id"));
+        }
 
         const result = await User.destroy({
             where: {
